requirements.txt
langchain==0.0.267
langchain-community==0.0.13
langchain-google-genai==0.0.5
flask==2.3.3
flask-socketio==5.3.5
PyPDF2==3.0.1
python-dotenv==1.0.0
cartesia